# Fitervari_Backend/Swagger API-Spezifikation/swagger-fv.yaml
swagger: "2.0"
info:
  description: "Diplomarbeit"
  version: "1.0.0"
  title: "Fitervari"
host: "todo.com"
basePath: "/api"
tags:
- name: "app"
- name: "webapp"
schemes:
- "https"
- "http"
paths:
  /workout:
    get:
    put:
    delete:
  /workoutPlans:
    get:
      tags:
      - "app"
      - "webapp"
      summary: "Get plans by user"
      parameters:
      - in: "query"
        name: "user"
        type: "integer"
        required: true
      responses:
        "200":
          description: "Ok"
        "404":
          description: "Not found"
    put:
      tags:
      - "webapp"
      summary: "Add/edit workout plan"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "plan"
        description: "New workout plan"
        required: true
        schema:
          $ref: "#/definitions/WorkoutPlan"
      responses:
        "200":
          description: "Updated"
        "201":
          description: "Created"
        "400":
          description: "Invalid input"
  /workoutPlans/{id}:
    get:
      tags:
      - "app"
      - "webapp"
      summary: "Get workout plan"
      parameters:
      - in: "path"
        name: "id"
        type: "integer"
        description: "Id of the plan"
        required: true
      responses:
        "200":
          description: "Ok"
        "404":
          description: "Not found"
    delete:
      tags:
      - "webapp"
      summary: "Delete a plan"
      parameters:
      - in: "path"
        name: "id"
        type: "integer"
        required: true
      responses:
        "200":
          description: "Ok"
        "404":
          description: "Not found"
  /users:
    get:
      tags:
      - "webapp"
      summary: "Get all users"
      responses:
        "200":
          description: "Ok"
    put:
      tags:
      - "app"
      summary: "Update user"
      responses:
        "200":
          description: "Ok"
        "404":
          description: "Not found"
  /users/{id}:
    get:
      tags:
      - "app"
      - "webapp"
      summary: "Get user by id"
      parameters:
      - in: "path"
        name: "id"
        type: "integer"
        required: true
      responses:
        "200":
          description: "Ok"
        "404":
          description: "Not found"
  /token:
    get:
      tags:
      - "webapp"
      summary: "Get new login token for user"
      parameters:
      - in: "query"
        name: "user"
        type: "integer"
        required: true
      responses:
        "200":
          description: "Ok"
        "404":
          description: "Not found"
definitions:
  WorkoutPlan:
    type: "object"